feat: Add For sequence for repeating a set of sequences
This commit adds a new sequence, For, to the project. The For sequence allows for the repetition of a set of sequences a specified number of times. It takes in a count resolver and a sequences resolver, resolving them to determine the number of repetitions and the sequences to be played. This addition enhances the sequencing capabilities of the game, providing a way to easily repeat sequences.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.