feat: Update Equipment.cs to display a more descriptive message when equipping an accessory
This commit updates the Equipment.cs file to display a more descriptive message when equipping an accessory. The message "使えそうだ。空いてる装飾品枠に装備した。" has been changed to "使えそうだ。空いてるアクセサリー枠に装備した。" to provide clearer information to the player.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.