feat: Add onBeginTurn sequence to BattleCharacterSequences.cs
Add a new method, PlayOnBeginTurnAsync, to the BattleCharacterSequences class. This method plays a sequence of actions when a battle character's turn begins. The sequence is defined by the onBeginTurn ScriptableSequences. This addition enhances the battle system by allowing specific actions to be performed at the start of a character's turn.