feat: Add ChangeWeapon command invoker
The code changes include adding a new file `ChangeWeapon.cs` to the `CommandInvokers` directory. This file contains the implementation of the `ChangeWeapon` class, which is responsible for changing the weapon of a battle character. The class has a constructor that takes two parameters: `equipmentWeaponInventoryIndex` and `changeWeaponId`. It also implements the `ICommandInvoker` interface and provides the necessary methods.

Resolves: #123