feat: Add PlayAsync method to ScreenEffectView
The code changes include adding a new method `PlayAsync` to the `ScreenEffectView` class. This method allows for playing a specific animation asynchronously by accessing the `HKUIDocument` and `SequenceMonobehaviour` components. This addition enhances the functionality of the `ScreenEffectView` by providing a convenient way to play animations with the specified animation name.

Resolves: #123