feat: Add PlayScreenEffectAnimation sequence class
The code changes include adding a new sequence class `PlayScreenEffectAnimation` to the `Sequences` directory. This class implements the `ISequence` interface and provides a `PlayAsync` method that plays a screen effect animation asynchronously. The animation name is specified through a serialized field. This addition enhances the functionality of the game by allowing for the execution of screen effect animations during gameplay.