feat: Add RegisterContainerBattleCharacterTurnCount sequence
This commit adds a new sequence called `RegisterContainerBattleCharacterTurnCount` to the project. This sequence is responsible for registering the turn count of a battle character in a container. It takes into account the target type (self or target) and resolves the key using a string resolver. This sequence will be used to track the turn count of battle characters during gameplay.