feat: Fix loading issue for item thumbnails
The code changes in `MasterData.cs` address a loading issue for item thumbnails. Previously, if the `ThumbnailId` was not empty but the corresponding `Thumbnail` was null, a warning message was logged. This commit fixes the issue by adding a check to ensure that the `ThumbnailId` is not empty before attempting to load the thumbnail.