feat: Fix bug in save data sync
The code changes fix a bug in the sync of save data. This bug caused the isValid flag in the suspendData object to not be properly updated, resulting in incorrect behavior when resuming from a saved game. This fix ensures that the isValid flag is correctly set to false after syncing the save data, and a confirmation dialog is shown to inform the player that the saved data has been deleted.