feat: Add SetIsTrueEndRoute sequence class to FloorEvent database
This commit adds a new sequence class called SetIsTrueEndRoute to the FloorEvent database. This class is responsible for setting the isTrueEndRoute property to a specified boolean value. The value is resolved using a BooleanResolver/Constant from the UnitySequencerSystem.Resolvers namespace. The default value is set to 0.

The SetIsTrueEndRoute sequence class is referenced in the FloorEvent assets with the following RID: 6455563984491249773.

This commit also includes changes to the FloorEvent assets 102003.1.0.asset and 102003.2.0.asset, adding the SetIsTrueEndRoute sequence class and its associated BooleanResolver/Constant valueResolver.

Refs: #issue-number